Add separate files for the different types of button components to help split apart the larger DMScreen component
